requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==45.0.7
defusedxml==0.7.1
django==5.2.6
django-allauth==65.12.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
djoser==2.3.3
drf-nested-routers==0.95.0
idna==3.10
mysqlclient==2.2.7
oauthlib==3.3.1
pillow==11.3.0
pip==25.0.1
pycparser==2.23
PyJWT==2.10.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
simplejwt==2.0.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
typing==3.7.4.3
tzdata==2025.2
urllib3==2.5.0
utils==1.0.2
